build: separate include dir
