+++ configure.py
@@ -17,14 +17,19 @@ parser = argparse.ArgumentParser(
         formatter_class = Formatter
         )
 parser_deps = parser.add_argument_group('dependency installation paths')
-parser_deps.add_argument( '--hipo', default=SYSTEM_ASSUMPTION, type=str, help='path to `hipo` installation' )
-parser_deps.add_argument( '--fmt',  default=SYSTEM_ASSUMPTION, type=str, help='path to `fmt` installation'  )
-parser_build = parser.add_argument_group('iguana build settings')
-parser_build.add_argument( '--prefix', default='iguana',           type=str, help='iguana installation prefix'         )
-parser_build.add_argument( '--build',  default='build-iguana',     type=str, help='iguana buildsystem directory'       )
-parser_build.add_argument( '--ini',    default='build-iguana.ini', type=str, help='name of the output config INI file' )
-parser_build.add_argument( '--relocatable', default=True, action=argparse.BooleanOptionalAction, help='make installation relocatable or not')
+parser_deps.add_argument( '--hipo', default=SYSTEM_ASSUMPTION, type=str, help='path to `hipo` installation')
+parser_deps.add_argument( '--fmt', default=SYSTEM_ASSUMPTION, type=str, help='path to `fmt` installation')
+parser_build = parser.add_argument_group('build settings')
+parser_build.add_argument( '--prefix', default='iguana', type=str, help='iguana installation prefix')
+parser_build.add_argument( '--examples', default=False, action=argparse.BooleanOptionalAction, help='build examples or not')
+parser_build = parser.add_argument_group('advanced settings')
+parser_build.add_argument( '--build', default='build-iguana', type=str, help='iguana buildsystem directory')
+parser_build.add_argument( '--ini', default='build-iguana.ini', type=str, help='name of the output config INI file')
+parser_build.add_argument( '--relocatable', default=True, action=argparse.BooleanOptionalAction, help='make pkg-config `.pc` file relocatable or not')
 args = parser.parse_args()
+
+# get prefix absolute path
+prefix = os.path.realpath(args.prefix)
 
 # set dependency paths
 cmake_prefix_path = []
@@ -47,14 +52,16 @@ def meson_string_array(arr):
 config = ConfigParser(allow_no_value=True)
 config.add_section('built-in options')
 if(len(cmake_prefix_path) > 0):
-    config.set('built-in options', '; dependency paths for: ' + ','.join(cmake_deps))
+    config.set('built-in options', '; path to dependencies: ' + ','.join(cmake_deps))
     config.set('built-in options', 'cmake_prefix_path', meson_string_array(cmake_prefix_path))
 if(len(pkg_config_path) > 0):
-    config.set('built-in options', '; dependency paths for: ' + ','.join(pkg_config_deps))
+    config.set('built-in options', '; path to dependencies: ' + ','.join(pkg_config_deps))
     config.set('built-in options', 'pkg_config_path', meson_string_array(pkg_config_path))
 config.set('built-in options', '; installation settings')
-config.set('built-in options', 'prefix', f'\'{os.path.realpath(args.prefix)}\'')
+config.set('built-in options', 'prefix', f'\'{prefix}\'')
+config.set('built-in options', 'libdir', '\'lib\'') # make all systems use lib/
 config.set('built-in options', 'pkgconfig.relocatable', f'{args.relocatable}')
+config.set('built-in options', 'examples', f'{args.examples}')
 
 # write the INI file
 with open(args.ini, 'w') as fp:
