--- examples/build_with_meson/00-basic.cc
@@ -1,1 +1,0 @@
-../00-basic.cc
