+++ include/iguana/Object.h
@@ -3,7 +3,7 @@
 #include <string>
 #include <memory>
 
-#include "Logger.h"
+#include "iguana/Logger.h"
 
 namespace iguana {
