+++ meson.options
@@ -1,0 +1,1 @@
+option('examples', type: 'boolean', value: false, description: 'Build examples')
