+++ src/services/meson.build
@@ -1,9 +1,3 @@
-services_headers = [
-  'Logger.h',
-  'Object.h',
-  'Algorithm.h',
-]
-
 services_sources = [
   'Logger.cc',
   'Object.cc',
@@ -14,14 +8,8 @@ services_lib = shared_library(
   'IguanaServices',
   services_sources,
   include_directories: project_inc,
-  dependencies:        [ fmt_dep, hipo_dep ],
+  dependencies:        project_deps,
   install:             true,
   install_rpath:       project_lib_rpath,
 )
-package_libs += services_lib
-
-services_dep = declare_dependency(
-  dependencies: [ fmt_dep, hipo_dep ]
-)
-
-install_headers(services_headers, subdir: meson.project_name())
+project_libs += services_lib
